The probabilites fit in 16 bits, so make the context array smaller
git-svn-id: svn://svn.mplayerhq.hu/soc/dirac@3928 66570020-b32e-0410-962d-e44e8f9afd43

+++ libavcodec/dirac_arith.h
@@ -86,7 +86,7 @@ typedef struct dirac_arith_state {
     const uint8_t *bytestream;
     const uint8_t *bytestream_end;
 
-    unsigned int contexts[ARITH_CONTEXT_COUNT];
+    uint16_t contexts[ARITH_CONTEXT_COUNT];
 
     PutBitContext *pb;
 } dirac_arith_state;
